+++ hephaistos-data/lua/Filters/EventPresentation.lua
@@ -1,4 +1,4 @@
--- vignettes flashes when getting hit
+-- vignette flashes when getting hit
 Hephaistos.CreateAnimation[HeroDamagePresentationThread] = function(params)
 	return Hephaistos.MatchAll(params, { Name = "BloodFrame", UseScreenLocation = true, OffsetX = ScreenCenterX, OffsetY = ScreenCenterY, GroupName = "Vignette" })
 end