+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 -   Fix locked keepsakes icons.
+-   Fix fire animation when getting by lava.
 
 ## [v0.8.0] - 2021-08-19